move formatGooglePlacesResultToAddress to shared folder
